Move some more code from root build.gradle.kts into buildSrc
