+++ build.gradle.kts
@@ -1,18 +1,5 @@
 
-import com.soywiz.korge.gradle.targets.*
-import com.soywiz.korge.gradle.targets.android.*
-import com.soywiz.korge.gradle.targets.ios.*
-import com.soywiz.korge.gradle.targets.jvm.*
-import com.soywiz.korge.gradle.targets.native.*
-import com.soywiz.korge.gradle.util.*
-import com.soywiz.korge.gradle.util.create
-import com.soywiz.korlibs.*
-import com.soywiz.korlibs.modules.*
-import com.soywiz.korlibs.modules.KorgeJavaExec
-import org.gradle.kotlin.dsl.kotlin
-import org.jetbrains.kotlin.gradle.tasks.*
-import java.net.*
-import java.nio.file.*
+import com.soywiz.korlibs.root.*
 
 buildscript {
     val kotlinVersion: String = libs.versions.kotlin.get()
@@ -45,1003 +32,8 @@ buildscript {
 plugins {
 	java
     kotlin("multiplatform")
-    id("org.jetbrains.dokka") version "1.7.10" apply true
     signing
     `maven-publish`
 }
 
-val kotlinVersion: String = project.kotlinVersion
-val realKotlinVersion = project.realKotlinVersion
-val nodeVersion: String = project.nodeVersion
-val androidBuildGradleVersion: String = project.androidBuildGradleVersion
-val forcedVersion = System.getenv("FORCED_VERSION")
-
-RootKorlibsPlugin(project)
-
-subprojects {
-    val doConfigure = mustAutoconfigureKMM()
-
-    if (doConfigure) {
-        val isSample = project.isSample
-        val hasAndroid = doEnableKotlinAndroid && hasAndroidSdk
-        //val hasAndroid = !isSample && true
-        val mustPublish = !isSample
-
-        // AppData\Local\Android\Sdk\tools\bin>sdkmanager --licenses
-        apply(plugin = "kotlin-multiplatform")
-
-        if (hasAndroid) {
-            if (isSample) {
-                apply(plugin = "com.android.application")
-            } else {
-                apply(plugin = "com.android.library")
-            }
-            apply(from = "${rootProject.rootDir}/build.android.gradle")
-            if (isSample) {
-                apply(from = "${rootProject.rootDir}/build.android.application.gradle")
-            }
-        }
-
-        if (isSample && doEnableKotlinNative && isMacos) {
-            configureNativeIos()
-        }
-
-        if (!isSample && rootProject.plugins.hasPlugin("org.jetbrains.dokka")) {
-            apply(plugin = "org.jetbrains.dokka")
-
-            tasks.dokkaHtml.configure {
-                offlineMode.set(true)
-            }
-
-            //dokkaHtml {
-            //    // Used to prevent resolving package-lists online. When this option is set to true, only local files are resolved
-            //    offlineMode.set(true)
-            //}
-
-            tasks {
-                //val dokkaCopy by creating(Task::class) {
-                //    dependsOn("dokkaHtml")
-                //    doLast {
-                //        val ffrom = File(project.buildDir, "dokka/html")
-                //        val finto = File(project.rootProject.projectDir, "build/dokka-all/${project.name}")
-                //        copy {
-                //            from(ffrom)
-                //            into(finto)
-                //        }
-                //        File(finto, "index-redirect.html").writeText("<meta http-equiv=\"refresh\" content=\"0; url=${project.name}\">\n")
-                //    }
-                //}
-            }
-        }
-
-        if (mustPublish) {
-            apply(plugin = "maven-publish")
-        }
-
-        tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
-            kotlinOptions.suppressWarnings = true
-        }
-
-        afterEvaluate {
-            val jvmTest = tasks.findByName("jvmTest")
-            if (jvmTest is org.jetbrains.kotlin.gradle.targets.jvm.tasks.KotlinJvmTest) {
-                val jvmTestFix = tasks.create("jvmTestFix", Test::class) {
-                    group = "verification"
-                    environment("UPDATE_TEST_REF", "true")
-                    testClassesDirs = jvmTest.testClassesDirs
-                    classpath = jvmTest.classpath
-                    bootstrapClasspath = jvmTest.bootstrapClasspath
-                    if (!JvmAddOpens.beforeJava9) jvmArgs(*JvmAddOpens.createAddOpensTypedArray())
-                    if (headlessTests) systemProperty("java.awt.headless", "true")
-                }
-                if (!JvmAddOpens.beforeJava9) jvmTest.jvmArgs(*JvmAddOpens.createAddOpensTypedArray())
-                if (headlessTests) jvmTest.systemProperty("java.awt.headless", "true")
-            }
-        }
-
-        kotlin {
-            //explicitApi()
-            //explicitApiWarning()
-
-            metadata {
-                compilations.all {
-                    kotlinOptions.suppressWarnings = true
-                }
-            }
-            jvm {
-                compilations.all {
-                    kotlinOptions.jvmTarget = "1.8"
-                    kotlinOptions.suppressWarnings = true
-                    kotlinOptions.freeCompilerArgs = listOf("-Xno-param-assertions")
-                    //kotlinOptions.
-
-                    // @TODO:
-                    // Tested on Kotlin 1.4.30:
-                    // Class org.luaj.vm2.WeakTableTest.WeakKeyTableTest
-                    // java.lang.AssertionError: expected:<null> but was:<mydata-111>
-                    //kotlinOptions.useIR = true
-                }
-            }
-            js(org.jetbrains.kotlin.gradle.plugin.KotlinJsCompilerType.IR) {
-                browser {
-                    compilations.all {
-                        //kotlinOptions.sourceMap = true
-                        kotlinOptions.suppressWarnings = true
-                    }
-                    testTask {
-                        useKarma {
-                            useChromeHeadless()
-                            useConfigDirectory(File(rootDir, "karma.config.d"))
-                        }
-                    }
-                }
-                nodejs {
-                    testTask {
-                        useMocha()
-                    }
-                }
-            }
-            if (hasAndroid) {
-                apply(from = "${rootProject.rootDir}/build.android.srcset.gradle")
-            }
-
-            val desktopAndMobileTargets = ArrayList<org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget>().apply {
-                if (doEnableKotlinNative) addAll(nativeTargets(project))
-                if (doEnableKotlinMobile) addAll(mobileTargets(project))
-            }.toList()
-
-            for (target in desktopAndMobileTargets) {
-                target.compilations.all {
-                    // https://github.com/JetBrains/kotlin/blob/ec6c25ef7ee3e9d89bf9a03c01e4dd91789000f5/kotlin-native/konan/konan.properties#L875
-                    kotlinOptions.freeCompilerArgs = ArrayList<String>().apply {
-                        // Raspberry Pi doesn't support mimalloc at this time
-                        if (useMimalloc && !target.name.contains("Arm32Hfp")) add("-Xallocator=mimalloc")
-                        add("-Xoverride-konan-properties=clangFlags.mingw_x64=-cc1 -emit-obj -disable-llvm-passes -x ir -target-cpu x86-64")
-                    }
-                    kotlinOptions.freeCompilerArgs += listOf(
-                        "-Xbinary=memoryModel=experimental",
-                        // @TODO: https://youtrack.jetbrains.com/issue/KT-49234#focus=Comments-27-5293935.0-0
-                        //"-Xdisable-phases=RemoveRedundantCallsToFileInitializersPhase",
-                    )
-                    kotlinOptions.suppressWarnings = true
-                }
-            }
-
-            // common
-            //    js
-            //    concurrent // non-js
-            //      jvmAndroid
-            //         android
-            //         jvm
-            //      native
-            //         kotlin-native
-            //    nonNative: [js, jvmAndroid]
-            sourceSets {
-
-                data class PairSourceSet(val main: org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet, val test: org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet) {
-                    fun get(test: Boolean) = if (test) this.test else this.main
-                    fun dependsOn(other: PairSourceSet) {
-                        main.dependsOn(other.main)
-                        test.dependsOn(other.test)
-                    }
-                }
-
-                //fun createPairSourceSet(name: String, vararg dependencies: PairSourceSet, block: org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet.(test: Boolean) -> Unit = { }): PairSourceSet {
-                fun createPairSourceSet(name: String, dependency: PairSourceSet? = null, block: org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet.(test: Boolean) -> Unit = { }): PairSourceSet { val dependencies = listOfNotNull(dependency)
-                    //println("${project.name}: CREATED SOURCE SET: \"${name}Main\"")
-                    val main = maybeCreate("${name}Main").apply { block(false) }
-                    val test = maybeCreate("${name}Test").apply { block(true) }
-                    return PairSourceSet(main, test).also {
-                        for (dependency in dependencies) {
-                            it.dependsOn(dependency)
-                        }
-                    }
-                }
-
-                val common = createPairSourceSet("common") { test ->
-                    dependencies {
-                        if (test) {
-                            implementation(kotlin("test-common"))
-                            implementation(kotlin("test-annotations-common"))
-                        } else {
-                            implementation(kotlin("stdlib-common"))
-                        }
-                    }
-                }
-
-                val concurrent = createPairSourceSet("concurrent", common)
-                val jvmAndroid = createPairSourceSet("jvmAndroid", concurrent)
-
-                // Default source set for JVM-specific sources and dependencies:
-                // JVM-specific tests and their dependencies:
-                val jvm = createPairSourceSet("jvm", jvmAndroid) { test ->
-                    dependencies {
-                        if (test) {
-                            implementation(kotlin("test-junit"))
-                        } else {
-                            implementation(kotlin("stdlib-jdk8"))
-                        }
-                    }
-                }
-
-                if (hasAndroid) {
-                    val android = createPairSourceSet("android", jvmAndroid) { test ->
-                        dependencies {
-                            if (test) {
-                                //implementation(kotlin("test"))
-                                //implementation(kotlin("test-junit"))
-                                implementation(kotlin("test-junit"))
-                            } else {
-                                //implementation(kotlin("stdlib"))
-                                //implementation(kotlin("stdlib-jdk8"))
-                            }
-                        }
-                    }
-                }
-
-                val js = createPairSourceSet("js", common) { test ->
-                    dependencies {
-                        if (test) {
-                            implementation(kotlin("test-js"))
-                        } else {
-                            implementation(kotlin("stdlib-js"))
-                        }
-                    }
-                }
-
-                if (doEnableKotlinNative) {
-                    val native by lazy { createPairSourceSet("native", concurrent) }
-                    val posix by lazy { createPairSourceSet("posix", native) }
-                    val darwin by lazy { createPairSourceSet("darwin", posix) }
-
-                    val linux by lazy { createPairSourceSet("linux", posix) }
-                    val macos by lazy { createPairSourceSet("macos", darwin) }
-                    val mingw by lazy { createPairSourceSet("mingw", native) }
-
-                    val nativeTargets = nativeTargets(project)
-
-                    for (target in nativeTargets) {
-                        val native = createPairSourceSet(target.name)
-                        when {
-                            target.isWin -> native.dependsOn(mingw)
-                            target.isMacos -> native.dependsOn(macos)
-                            target.isLinux -> native.dependsOn(linux)
-                        }
-                    }
-
-                    val darwinMobile by lazy { createPairSourceSet("darwinMobile", darwin) }
-                    val iosTvos by lazy { createPairSourceSet("iosTvos", darwinMobile) }
-                    val watchos by lazy { createPairSourceSet("watchos", darwinMobile) }
-                    val tvos by lazy { createPairSourceSet("tvos", iosTvos) }
-                    val ios by lazy { createPairSourceSet("ios", iosTvos) }
-
-                    for (target in mobileTargets(project)) {
-                        val native = createPairSourceSet(target.name)
-                        when {
-                            target.isIos -> native.dependsOn(ios)
-                            target.isWatchos -> native.dependsOn(watchos)
-                            target.isTvos -> native.dependsOn(tvos)
-                        }
-                    }
-
-                    /*
-                    for (baseName in listOf(
-                        "nativeInteropMain",
-                        "posixInteropMain",
-                        "darwinInteropMain",
-                        "linuxInteropMain",
-                    )) {
-                        val nativeInteropMainFolder = file("src/$baseName/kotlin")
-                        if (nativeInteropMainFolder.isDirectory) {
-                            val currentNativeTarget = currentPlatformNativeTarget(project)
-                            // @TODO: Copy instead of use the same source folder
-                            for (target in allNativeTargets(project)) {
-                                if (baseName.contains("posix", ignoreCase = true) && !target.isPosix) continue
-                                if (baseName.contains("darwin", ignoreCase = true) && !target.isApple) continue
-                                if (baseName.contains("linux", ignoreCase = true) && !target.isLinux) continue
-
-                                val sourceSet = this@sourceSets.maybeCreate("${target.name}Main")
-                                val folder = when {
-                                    target == currentNativeTarget -> nativeInteropMainFolder
-                                    else -> {
-                                        file("build/${baseName}Copy${target.name}").also { outFolder ->
-                                            outFolder.mkdirs()
-                                            sync {
-                                                from(nativeInteropMainFolder)
-                                                into(outFolder)
-                                            }
-                                        }
-                                    }
-                                }
-                                sourceSet.kotlin.srcDir(folder)
-                            }
-                        }
-                    }
-                    */
-
-                    // Copy test resources
-                    afterEvaluate {
-                        for (targetV in (nativeTargets + listOf(iosX64(), iosSimulatorArm64()))) {
-                            val target = targetV.name
-                            val taskName = "copyResourcesToExecutable_$target"
-                            val targetTestTask = tasks.findByName("${target}Test") as? org.jetbrains.kotlin.gradle.targets.native.tasks.KotlinNativeTest? ?: continue
-                            val compileTestTask = tasks.findByName("compileTestKotlin${target.capitalize()}") ?: continue
-                            val compileMainTask = tasks.findByName("compileKotlin${target.capitalize()}") ?: continue
-
-                            //println("$targetTestTask -> $target")
-
-                            tasks {
-                                create<Copy>(taskName) {
-                                    for (sourceSet in kotlin.sourceSets) {
-                                        from(sourceSet.resources)
-                                    }
-
-                                    into(targetTestTask.executable.parentFile)
-                                }
-                            }
-
-                            targetTestTask.inputs.files(
-                                *compileTestTask.outputs.files.files.toTypedArray(),
-                                *compileMainTask.outputs.files.files.toTypedArray()
-                            )
-
-                            targetTestTask.dependsOn(taskName)
-                            //println(".target=$target")
-                            if (target == "mingwX64") {
-                                afterEvaluate {
-                                    afterEvaluate {
-                                        tasks.findByName("mingwX64TestWine")?.let {
-                                            //println("***************++")
-                                            it?.dependsOn(taskName)
-                                        }
-                                    }
-                                }
-                            }
-                            if (target == "linuxX64") {
-                                afterEvaluate {
-                                    afterEvaluate {
-                                        tasks.findByName("linuxX64TestLima")?.let {
-                                            //println("***************++")
-                                            it?.dependsOn(taskName)
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-internal var _webServer: DecoratedHttpServer? = null
-
-samples {
-    // @TODO: Patch, because runDebugReleaseExecutableMacosArm64 is not created!
-    if (isMacos && isArm && doEnableKotlinNative) {
-        project.tasks {
-            afterEvaluate {
-                for (kind in listOf("Debug", "Release")) {
-                    val linkTaskName = "link${kind}ExecutableMacosArm64"
-                    val runTaskName = "run${kind}ExecutableMacosArm64"
-                    val linkExecutableMacosArm64 = project.tasks.findByName(linkTaskName) as org.jetbrains.kotlin.gradle.tasks.KotlinNativeLink
-                    if (project.tasks.findByName(runTaskName) == null) {
-                        val runExecutableMacosArm64 = project.tasks.create(runTaskName, Exec::class) {
-                            dependsOn(linkExecutableMacosArm64)
-                            group = "run"
-                            commandLine(linkExecutableMacosArm64.binary.outputFile)
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // @TODO: Move to KorGE plugin
-    project.tasks {
-        // https://www.baeldung.com/java-instrumentation
-        val runJvm by creating(KorgeJavaExec::class) {
-            group = "run"
-            mainClass.set("MainKt")
-        }
-
-        val timeBeforeCompilationFile = File(project.buildDir, "timeBeforeCompilation")
-        val httpPort = 22011
-
-        val compileKotlinJvmAndNotifyBefore by creating(Task::class) {
-            doFirst {
-                KorgeReloadNotifier.beforeBuild(timeBeforeCompilationFile)
-            }
-        }
-        afterEvaluate {
-            tasks.findByName("compileKotlinJvm")?.mustRunAfter("compileKotlinJvmAndNotifyBefore")
-        }
-        val compileKotlinJvmAndNotify by creating(Task::class) {
-            dependsOn("compileKotlinJvmAndNotifyBefore", "compileKotlinJvm")
-            doFirst {
-                KorgeReloadNotifier.afterBuild(timeBeforeCompilationFile, httpPort)
-            }
-        }
-        for (enableRedefinition in listOf(false, true)) {
-            val taskName = when (enableRedefinition) {
-                false -> "runJvmAutoreload"
-                true -> "runJvmAutoreloadWithRedefinition"
-            }
-            create(taskName, KorgeJavaExec::class) {
-                dependsOn(":korge-reload-agent:jar", "compileKotlinJvm")
-                group = "run"
-                mainClass.set("MainKt")
-                afterEvaluate {
-                    val agentJarTask: Jar = project(":korge-reload-agent").tasks.findByName("jar") as Jar
-                    val outputJar = agentJarTask.outputs.files.files.first()
-                    //println("agentJarTask=$outputJar")
-                    val compileKotlinJvm = tasks.findByName("compileKotlinJvm") as org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-                    val args = compileKotlinJvm.outputs.files.toList().joinToString(":::") { it.absolutePath }
-                    //val gradlewCommand = if (isWindows) "gradlew.bat" else "gradlew"
-                    //val continuousCommand = "${rootProject.rootDir}/$gradlewCommand --no-daemon --warn --project-dir=${rootProject.rootDir} --configuration-cache -t ${project.path}:compileKotlinJvmAndNotify"
-                    val continuousCommand =
-                        "-classpath ${rootProject.rootDir}/gradle/wrapper/gradle-wrapper.jar org.gradle.wrapper.GradleWrapperMain --no-daemon --warn --project-dir=${rootProject.rootDir} --configuration-cache -t ${project.path}:compileKotlinJvmAndNotify"
-                    jvmArgs("-javaagent:$outputJar=$httpPort:::$continuousCommand:::$enableRedefinition:::$args")
-                    environment("KORGE_AUTORELOAD", "true")
-                }
-            }
-        }
-
-        // esbuild
-        run {
-            fun runServer(blocking: Boolean) {
-                if (_webServer == null) {
-                    val address = "0.0.0.0"
-                    val port = 8080
-                    val server = staticHttpServer(File(project.buildDir, "www"), address = address, port = port)
-                    _webServer = server
-                    try {
-                        val openAddress = when (address) {
-                            "0.0.0.0" -> "127.0.0.1"
-                            else -> address
-                        }
-                        openBrowser("http://$openAddress:${server.port}/index.html")
-                        if (blocking) {
-                            while (true) {
-                                Thread.sleep(1000L)
-                            }
-                        }
-                    } finally {
-                        if (blocking) {
-                            println("Stopping web server...")
-                            server.server.stop(0)
-                            _webServer = null
-                        }
-                    }
-                }
-                _webServer?.updateVersion?.incrementAndGet()
-            }
-
-            val userGradleFolder = File(System.getProperty("user.home"), ".gradle")
-
-            val esbuildVersion = "0.12.22"
-            val wwwFolder = File(buildDir, "www")
-
-            val esbuildFolder = File(if (userGradleFolder.isDirectory) userGradleFolder else rootProject.buildDir, "esbuild")
-            val isWindows = org.apache.tools.ant.taskdefs.condition.Os.isFamily(org.apache.tools.ant.taskdefs.condition.Os.FAMILY_WINDOWS)
-            val esbuildCmdUnix = File(esbuildFolder, "bin/esbuild")
-            val esbuildCmdCheck = if (isWindows) File(esbuildFolder, "esbuild.cmd") else esbuildCmdUnix
-            val esbuildCmd = if (isWindows) File(esbuildFolder, "node_modules/esbuild/esbuild.exe") else esbuildCmdUnix
-
-            val npmInstallEsbuild = "npmInstallEsbuild"
-
-            val env by lazy { org.jetbrains.kotlin.gradle.targets.js.nodejs.NodeJsRootPlugin.apply(project.rootProject).requireConfigured() }
-            val ENV_PATH by lazy {
-                val NODE_PATH = File(env.nodeExecutable).parent
-                val PATH_SEPARATOR = File.pathSeparator
-                val OLD_PATH = System.getenv("PATH")
-                "$NODE_PATH$PATH_SEPARATOR$OLD_PATH"
-            }
-
-            if (rootProject.tasks.findByName(npmInstallEsbuild) == null) {
-                rootProject.tasks.create(npmInstallEsbuild, Exec::class) {
-                    dependsOn("kotlinNodeJsSetup")
-                    onlyIf { !esbuildCmdCheck.exists() && !esbuildCmd.exists() }
-
-                    val esbuildVersion = esbuildVersion
-                    doFirst {
-                        val npmCmd = arrayOf(
-                            File(env.nodeExecutable),
-                            File(env.nodeDir, "lib/node_modules/npm/bin/npm-cli.js").takeIf { it.exists() }
-                                ?: File(env.nodeDir, "node_modules/npm/bin/npm-cli.js").takeIf { it.exists() }
-                                ?: error("Can't find npm-cli.js in ${env.nodeDir} standard folders")
-                        )
-
-                        environment("PATH", ENV_PATH)
-                        commandLine(*npmCmd, "-g", "install", "esbuild@$esbuildVersion", "--prefix", esbuildFolder, "--scripts-prepend-node-path", "true")
-                    }
-                }
-            }
-
-            afterEvaluate {
-                val browserEsbuildResources by tasks.creating(Copy::class) {
-                    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-                    from(project.tasks.getByName("jsProcessResources").outputs.files)
-                    afterEvaluate {
-                        project.tasks.findByName("korgeProcessedResourcesJsMain")?.outputs?.files?.let {
-                            from(it)
-                        }
-                    }
-                    //for (sourceSet in gkotlin.js().compilations.flatMap { it.kotlinSourceSets }) from(sourceSet.resources)
-                    into(wwwFolder)
-                }
-                val compileDevelopmentExecutableKotlinJs = "compileDevelopmentExecutableKotlinJs"
-                val runJs by creating(Exec::class) {
-                    group = "run"
-                    //dependsOn("jsBrowserDevelopmentRun")
-                    dependsOn(browserEsbuildResources)
-                    dependsOn("::$npmInstallEsbuild")
-                    dependsOn(compileDevelopmentExecutableKotlinJs)
-                    //task.dependsOn(browserPrepareEsbuild)
-
-                    val jsPath = project.tasks.getByName(compileDevelopmentExecutableKotlinJs).outputs.files.first {
-                        it.extension.toLowerCase() == "js"
-                    }
-
-                    val output = File(wwwFolder, "${project.name}.js")
-                    inputs.file(jsPath)
-                    outputs.file(output)
-                    //task.environment("PATH", ENV_PATH)
-                    commandLine(ArrayList<Any>().apply {
-                        add(esbuildCmd)
-                        //add("--watch",)
-                        add("--bundle")
-                        //add("--minify")
-                        //add("--sourcemap=external")
-                        add(jsPath)
-                        add("--outfile=$output")
-                        // @TODO: Close this command on CTRL+C
-                        //if (run) add("--servedir=$wwwFolder")
-                    })
-
-                    doLast {
-                        runServer(!project.gradle.startParameter.isContinuous)
-                    }
-                }
-            }
-        }
-
-        val runJsWebpack by creating {
-            group = "run"
-            //dependsOn("jsBrowserDevelopmentRun")
-            dependsOn("jsBrowserProductionRun")
-        }
-
-        fun runNativeTaskNameWin(kind: String): String {
-            return "run${kind}ExecutableMingwX64"
-        }
-
-        fun runNativeTaskName(kind: String): String {
-            return when {
-                isWindows -> runNativeTaskNameWin(kind)
-                isMacos -> if (isArm) "run${kind}ExecutableMacosArm64" else "run${kind}ExecutableMacosX64"
-                else -> "run${kind}ExecutableLinuxX64"
-            }
-        }
-
-        fun Task.dependsOnNativeTask(kind: String) {
-            dependsOn(runNativeTaskName(kind))
-        }
-        val runNativeDebug by creating {
-            group = "run"
-            dependsOnNativeTask("Debug")
-        }
-        val runNativeRelease by creating {
-            group = "run"
-            dependsOnNativeTask("Release")
-        }
-        if (!isWindows) {
-            afterEvaluate {
-                for (type in CrossExecType.VALID_LIST) {
-                    for (deb in listOf("Debug", "Release")) {
-                        val linkTask = project.tasks.findByName("link${deb}Executable${type.nameWithArchCapital}") as? KotlinNativeLink? ?: continue
-                        tasks.create("runNative${deb}${type.interpCapital}", Exec::class.java) {
-                            group = "run"
-                            dependsOn(linkTask)
-                            commandLineCross(linkTask.binary.outputFile.absolutePath, type = type)
-                            this.environment("WINEDEBUG", "-all")
-                            workingDir = linkTask.binary.outputDirectory
-                        }
-                    }
-                }
-            }
-        }
-
-        //val jsRun by creating { dependsOn("jsBrowserDevelopmentRun") } // Already available
-        //val jvmRun by creating {
-        //    group = "run"
-        //    dependsOn(runJvm)
-        //}
-        //val run by getting(JavaExec::class)
-
-        //val processResources by getting {
-        //	dependsOn(processResourcesKorge)
-        //}
-    }
-
-    kotlin {
-        jvm {
-        }
-        js(org.jetbrains.kotlin.gradle.plugin.KotlinJsCompilerType.IR) {
-            browser {
-                binaries.executable()
-            }
-        }
-
-        tasks.getByName("jsProcessResources", Task::class).apply {
-            //println(this.outputs.files.toList())
-            doLast {
-                val targetDir = this.outputs.files.first()
-                val jsMainCompilation = kotlin.js(org.jetbrains.kotlin.gradle.plugin.KotlinJsCompilerType.IR).compilations["main"]!!
-
-                // @TODO: How to get the actual .js file generated/served?
-                val jsFile = File("${project.name}.js").name
-                val resourcesFolders = jsMainCompilation.allKotlinSourceSets
-                    .flatMap { it.resources.srcDirs } + listOf(File(rootProject.rootDir, "_template"))
-                //println("jsFile: $jsFile")
-                //println("resourcesFolders: $resourcesFolders")
-                fun readTextFile(name: String): String {
-                    for (folder in resourcesFolders) {
-                        val file = File(folder, name)?.takeIf { it.exists() } ?: continue
-                        return file.readText()
-                    }
-                    return ClassLoader.getSystemResourceAsStream(name)?.readBytes()?.toString(Charsets.UTF_8)
-                        ?: error("We cannot find suitable '$name'")
-                }
-
-                val indexTemplateHtml = readTextFile("index.v2.template.html")
-                val customCss = readTextFile("custom-styles.template.css")
-                val customHtmlHead = readTextFile("custom-html-head.template.html")
-                val customHtmlBody = readTextFile("custom-html-body.template.html")
-
-                //println(File(targetDir, "index.html"))
-
-                File(targetDir, "index.html").writeText(
-                    groovy.text.SimpleTemplateEngine().createTemplate(indexTemplateHtml).make(
-                        mapOf(
-                            "OUTPUT" to jsFile,
-                            //"TITLE" to korge.name,
-                            "TITLE" to "TODO",
-                            "CUSTOM_CSS" to customCss,
-                            "CUSTOM_HTML_HEAD" to customHtmlHead,
-                            "CUSTOM_HTML_BODY" to customHtmlBody
-                        )
-                    ).toString()
-                )
-            }
-        }
-
-        if (doEnableKotlinNative) {
-            for (target in nativeTargets(project)) {
-                target.apply {
-                    binaries {
-                        executable {
-                            entryPoint("entrypoint.main")
-                        }
-                    }
-                }
-            }
-
-            val nativeDesktopFolder = File(project.buildDir, "platforms/nativeDesktop")
-            //val nativeDesktopEntryPointSourceSet = kotlin.sourceSets.create("nativeDesktopEntryPoint")
-            //nativeDesktopEntryPointSourceSet.kotlin.srcDir(nativeDesktopFolder)
-            sourceSets.getByName("nativeMain") { kotlin.srcDir(nativeDesktopFolder) }
-
-            val createEntryPointAdaptorNativeDesktop = tasks.create("createEntryPointAdaptorNativeDesktop") {
-                val mainEntrypointFile = File(nativeDesktopFolder, "entrypoint/main.kt")
-
-                outputs.file(mainEntrypointFile)
-
-                // @TODO: Determine the package of the main file
-                doLast {
-                    mainEntrypointFile.also { it.parentFile.mkdirs() }.writeText("""
-                        package entrypoint
-
-                        import kotlinx.coroutines.*
-                        import main
-
-                        fun main(args: Array<String>) {
-                            runBlocking {
-                                main()
-                            }
-                        }
-                    """.trimIndent())
-                }
-            }
-
-            val nativeDesktopTargets = nativeTargets(project)
-            val allNativeTargets = nativeDesktopTargets
-
-            //for (target in nativeDesktopTargets) {
-                //target.compilations["main"].defaultSourceSet.dependsOn(nativeDesktopEntryPointSourceSet)
-            //    target.compilations["main"].defaultSourceSet.kotlin.srcDir(nativeDesktopFolder)
-            //}
-
-            for (target in allNativeTargets) {
-                for (binary in target.binaries) {
-                    val compilation = binary.compilation
-                    val copyResourcesTask = tasks.create("copyResources${target.name.capitalize()}${binary.name.capitalize()}", Copy::class) {
-                        dependsOn(getKorgeProcessResourcesTaskName(target, compilation))
-                        group = "resources"
-                        val isDebug = binary.buildType == org.jetbrains.kotlin.gradle.plugin.mpp.NativeBuildType.DEBUG
-                        val isTest = binary.outputKind == org.jetbrains.kotlin.gradle.plugin.mpp.NativeOutputKind.TEST
-                        val compilation = if (isTest) target.compilations["test"] else target.compilations["main"]
-                        //target.compilations.first().allKotlinSourceSets
-                        val sourceSet = compilation.defaultSourceSet
-                        from(sourceSet.resources)
-                        from(sourceSet.dependsOn.map { it.resources })
-                        into(binary.outputDirectory)
-                    }
-
-                    //compilation.compileKotlinTask.dependsOn(copyResourcesTask)
-                    binary.linkTask.dependsOn(copyResourcesTask)
-                    binary.compilation.compileKotlinTask.dependsOn(createEntryPointAdaptorNativeDesktop)
-                }
-            }
-        }
-    }
-
-    project.tasks {
-        val runJvm by getting(KorgeJavaExec::class)
-        //val prepareResourceProcessingClasses = create("prepareResourceProcessingClasses", Copy::class) {
-        //    dependsOn(jvmMainClasses)
-        //    afterEvaluate {
-        //        from(runJvm.korgeClassPath.toList().map { if (it.extension == "jar") zipTree(it) else it })
-        //    }
-        //    into(File(project.buildDir, "korgeProcessedResources/classes"))
-        //}
-
-        for (target in kotlin.targets) {
-            for (compilation in target.compilations) {
-                val processedResourcesFolder = File(project.buildDir, "korgeProcessedResources/${target.name}/${compilation.name}")
-                compilation.defaultSourceSet.resources.srcDir(processedResourcesFolder)
-                val korgeProcessedResources = create(getKorgeProcessResourcesTaskName(target, compilation)) {
-                    //dependsOn(prepareResourceProcessingClasses)
-                    dependsOn("jvmMainClasses")
-
-                    if (enabledSandboxResourceProcessor) {
-                        doLast {
-                            processedResourcesFolder.mkdirs()
-                            //URLClassLoader(prepareResourceProcessingClasses.outputs.files.toList().map { it.toURL() }.toTypedArray(), ClassLoader.getSystemClassLoader()).use { classLoader ->
-
-                            URLClassLoader(
-                                runJvm.korgeClassPath.toList().map { it.toURL() }.toTypedArray(),
-                                ClassLoader.getSystemClassLoader()
-                            ).use { classLoader ->
-                                val clazz = classLoader.loadClass("com.soywiz.korge.resources.ResourceProcessorRunner")
-                                val folders = compilation.allKotlinSourceSets.flatMap { it.resources.srcDirs }
-                                    .filter { it != processedResourcesFolder }.map { it.toString() }
-                                //println(folders)
-                                try {
-                                    clazz.methods.first { it.name == "run" }.invoke(
-                                        null,
-                                        classLoader,
-                                        folders,
-                                        processedResourcesFolder.toString(),
-                                        compilation.name
-                                    )
-                                } catch (e: java.lang.reflect.InvocationTargetException) {
-                                    val re = (e.targetException ?: e)
-                                    re.printStackTrace()
-                                    System.err.println(re.toString())
-                                }
-                            }
-                            System.gc()
-                        }
-                    }
-                }
-                //println(compilation.compileKotlinTask.name)
-                //println(compilation.compileKotlinTask.name)
-                //compilation.compileKotlinTask.finalizedBy(processResourcesKorge)
-                //println(compilation.compileKotlinTask)
-                //compilation.compileKotlinTask.dependsOn(processResourcesKorge)
-                if (compilation.compileKotlinTask.name != "compileKotlinJvm") {
-                    compilation.compileKotlinTask.dependsOn(korgeProcessedResources)
-                } else {
-                    compilation.compileKotlinTask.finalizedBy(korgeProcessedResources)
-                    getByName("runJvm").dependsOn(korgeProcessedResources)
-
-                }
-                //println(compilation.output.allOutputs.toList())
-                //println("$target - $compilation")
-
-            }
-        }
-    }
-}
-
-val projectVersion = project.version
-fun createBuildVersions(git: Boolean): String {
-    return """
-        package com.soywiz.korge.gradle
-        
-        object BuildVersions {
-            const val GIT = "${if (git) gitVersion else "main"}"
-            const val KOTLIN = "$realKotlinVersion"
-            const val NODE_JS = "$nodeVersion"
-            const val JNA = "${libs.versions.jna.get()}"
-            const val COROUTINES = "${libs.versions.kotlinx.coroutines.get()}"
-            const val ANDROID_BUILD = "$androidBuildGradleVersion"
-            const val KOTLIN_SERIALIZATION = "${libs.versions.kotlinx.serialization.get()}"
-            const val KRYPTO = "$projectVersion"
-            const val KLOCK = "$projectVersion"
-            const val KDS = "$projectVersion"
-            const val KMEM = "$projectVersion"
-            const val KORMA = "$projectVersion"
-            const val KORIO = "$projectVersion"
-            const val KORIM = "$projectVersion"
-            const val KORAU = "$projectVersion"
-            const val KORGW = "$projectVersion"
-            const val KORGE = "$projectVersion"
-        
-            val ALL_PROPERTIES by lazy { listOf(::GIT, ::KRYPTO, ::KLOCK, ::KDS, ::KMEM, ::KORMA, ::KORIO, ::KORIM, ::KORAU, ::KORGW, ::KORGE, ::KOTLIN, ::JNA, ::COROUTINES, ::ANDROID_BUILD, ::KOTLIN_SERIALIZATION) }
-            val ALL by lazy { ALL_PROPERTIES.associate { it.name to it.get() } }
-        }
-    """.trimIndent()
-}
-
-file("buildSrc/src/main/kotlinGen/com/soywiz/korge/gradle/BuildVersions.kt").writeTextIfChanged(createBuildVersions(git = false))
-file("korge-gradle-plugin/build/srcgen/com/soywiz/korge/gradle/BuildVersions.kt").writeTextIfChanged(createBuildVersions(git = true))
-
-
-subprojects {
-    afterEvaluate {
-        tasks {
-            val publishKotlinMultiplatformPublicationToMavenLocal = "publishKotlinMultiplatformPublicationToMavenLocal"
-            val publishKotlinMultiplatformPublicationToMavenRepository = "publishKotlinMultiplatformPublicationToMavenRepository"
-
-            val publishJvmLocal by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishJvmPublicationToMavenLocal")
-                    //dependsOn("publishMetadataPublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                } else if (findByName("publishToMavenLocal") != null) {
-                    dependsOn("publishToMavenLocal")
-                }
-            }
-
-            val publishJsLocal by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishJsPublicationToMavenLocal")
-                    //dependsOn("publishMetadataPublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                }
-            }
-
-            val publishMacosX64Local by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishMacosX64PublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                }
-            }
-            val publishMacosArm64Local by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishMacosArm64PublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                }
-            }
-            val publishIosX64Local by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishIosX64PublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                }
-            }
-            val publishIosArm64Local by creating(Task::class) {
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn("publishIosArm64PublicationToMavenLocal")
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                }
-            }
-            val publishMobileLocal by creating(Task::class) {
-                doFirst {
-                    //if (currentJavaVersion != 8) error("To use publishMobileRepo, must be used Java8, but used Java$currentJavaVersion")
-                }
-                run {
-                    val taskName = "publishJvmPublicationToMavenLocal"
-                    if (findByName(taskName) != null) {
-                        dependsOn(taskName)
-                    }
-                }
-                if (findByName(publishKotlinMultiplatformPublicationToMavenLocal) != null) {
-                    dependsOn(publishKotlinMultiplatformPublicationToMavenLocal)
-                    dependsOn("publishAndroidPublicationToMavenLocal")
-                    dependsOn("publishIosArm64PublicationToMavenLocal")
-                    dependsOn("publishIosX64PublicationToMavenLocal")
-                }
-            }
-
-            val customMavenUser = rootProject.findProperty("KORLIBS_CUSTOM_MAVEN_USER")?.toString()
-            val customMavenPass = rootProject.findProperty("KORLIBS_CUSTOM_MAVEN_PASS")?.toString()
-            val customMavenUrl = rootProject.findProperty("KORLIBS_CUSTOM_MAVEN_URL")?.toString()
-            val customPublishEnabled = forcedVersion != null
-                && !customMavenUser.isNullOrBlank()
-                && !customMavenPass.isNullOrBlank()
-                && !customMavenUrl.isNullOrBlank()
-
-            val publishMobileRepo by creating(Task::class) {
-                doFirst {
-                    if (currentJavaVersion != 8) {
-                        error("To use publishMobileRepo, must be used Java8, but used Java$currentJavaVersion")
-                    }
-                    if (!customPublishEnabled) {
-                        error("To use publishMobileRepo, must set `FORCED_VERSION=...` environment variable, and in ~/.gradle/gradle.properties : KORLIBS_CUSTOM_MAVEN_USER, KORLIBS_CUSTOM_MAVEN_PASS & KORLIBS_CUSTOM_MAVEN_URL")
-                    }
-                }
-                if (customPublishEnabled) {
-                    run {
-                        val taskName = "publishJvmPublicationToMavenRepository"
-                        if (findByName(taskName) != null) {
-                            dependsOn(taskName)
-                        }
-                    }
-                    if (findByName(publishKotlinMultiplatformPublicationToMavenRepository) != null) {
-                        dependsOn(publishKotlinMultiplatformPublicationToMavenRepository)
-                        dependsOn("publishAndroidPublicationToMavenRepository")
-                        dependsOn("publishIosArm64PublicationToMavenRepository")
-                        dependsOn("publishIosX64PublicationToMavenRepository")
-                    }
-                }
-            }
-        }
-        //tasks.withType(Test::class.java).all {
-        tasks.withType(AbstractTestTask::class.java).all {
-            testLogging {
-                //setEvents(setOf("passed", "skipped", "failed", "standardOut", "standardError"))
-                setEvents(setOf("skipped", "failed", "standardError"))
-                exceptionFormat = org.gradle.api.tasks.testing.logging.TestExceptionFormat.FULL
-            }
-        }
-        tasks {
-            afterEvaluate {
-                for (type in CrossExecType.VALID_LIST) {
-                    val linkDebugTest = project.tasks.findByName("linkDebugTest${type.nameWithArchCapital}") as? KotlinNativeLink?
-                    if (linkDebugTest != null) {
-                        tasks.create("${type.nameWithArch}Test${type.interpCapital}", com.soywiz.korge.gradle.targets.native.KotlinNativeCrossTest::class.java, Action {
-                            val link = linkDebugTest
-                            val testResultsDir = project.buildDir.resolve(TestingBasePlugin.TEST_RESULTS_DIR_NAME)
-                            val testReportsDir = project.extensions.getByType(ReportingExtension::class.java).baseDir.resolve(TestingBasePlugin.TESTS_DIR_NAME)
-                            //this.configureConventions()
-
-                            val htmlReport = org.gradle.api.internal.plugins.DslObject(reports.html)
-                            val xmlReport = org.gradle.api.internal.plugins.DslObject(reports.junitXml)
-                            xmlReport.conventionMapping.map("destination") { testResultsDir.resolve(name) }
-                            htmlReport.conventionMapping.map("destination") { testReportsDir.resolve(name) }
-
-                            this.type = type
-                            this.executable = link.binary.outputFile
-                            this.workingDir = link.binary.outputDirectory.absolutePath
-                            this.binaryResultsDirectory.set(testResultsDir.resolve("$name/binary"))
-                            this.environment("WINEDEBUG", "-all")
-                            group = "verification"
-                            dependsOn(link)
-                        })
-                    }
-                }
-            }
-        }
-    }
-}
-
-// https://youtrack.jetbrains.com/issue/KT-48273
-afterEvaluate {
-    rootProject.extensions.configure<org.jetbrains.kotlin.gradle.targets.js.nodejs.NodeJsRootExtension> {
-        versions.webpackDevServer.version = "4.0.0"
-    }
-}
-
-if (isLinux) {
-    project.logger.info("LD folders: ${LDLibraries.ldFolders}")
-    for (lib in listOf("libGL.so.1", "libopenal.so.1")) {
-        if (!LDLibraries.hasLibrary(lib)) {
-            System.err.println("Can't find $lib. Please: sudo apt-get -y install freeglut3 libopenal1")
-        }
-    }
-}
-
-//println("currentJavaVersion=${com.soywiz.korlibs.currentJavaVersion()}")
+RootKorlibsPlugin.doInit(rootProject)
