+++ buildSrc/build.gradle
@@ -15,6 +15,7 @@ if (System.getProperty("java.version").startsWith("1.8") || System.getProperty("
 
 dependencies {
     implementation("org.jetbrains.kotlinx:kover:0.6.1")
+    implementation("org.jetbrains.dokka:dokka-gradle-plugin:1.7.20")
     implementation("com.android.tools.build:gradle:$androidBuildGradleVersion")
     implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:${libs.versions.kotlin.get()}")
