+++ buildSrc/src/main/kotlin/com/soywiz/korge/gradle/util/DslExt.kt
@@ -1,7 +1,15 @@
 package com.soywiz.korge.gradle.util
 
+import kotlinx.kover.api.*
 import org.gradle.api.*
+import org.gradle.api.tasks.*
 import kotlin.reflect.*
+
+fun org.gradle.api.Project.`koverMerged`(configure: Action<KoverMergedConfig>): Unit = (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("koverMerged", configure)
+fun org.gradle.api.Project.`kover`(configure: Action<kotlinx.kover.api.KoverProjectConfig>): Unit = (this as org.gradle.api.plugins.ExtensionAware).extensions.configure("kover", configure)
+//fun TaskContainer.`dokkaHtml`(configure: org.jetbrains.dokka.gradle.DokkaTask.() -> Unit) {
+//    configure(named<org.jetbrains.dokka.gradle.DokkaTask>("dokkaHtml").get())
+//}
 
 /**
  * Retrieves the [ext][org.gradle.api.plugins.ExtraPropertiesExtension] extension.
